Expose Block#fromString
